// B_Ang/src/app/models/issue.ts
export class Issue {
  id: number;
  docid: number;
  docautorecive: string;
  scanningpath: string;
  realpolicynum: string;
  ispolicyactive: string;
  iscaseonwb: string;
  routingidtypelist: string;
  lineofbusiness: string;
  internalstatusvalid: string;
  internalstatusinvalid: string;
  casestatusvalid: string;
  casestatusinvalid: string;
  applicationtypevalid: string;
  applicationtypeinvalid: string;
  plan: string;
  replacementlist: string;
  fieldreporting: string;
  isunderwriterassigned: string;
  isunderwriteridvalid: string;
  unitcodevalid: string;
  unitcodeinvalid: string;
  notificationthrough: string;
  sharedworklistid: string;
  underwriterworkListid: string;
  processcoordinatormethod: string;
  documentroutingtype: string;
}